Finish refactoring quests into subdirectories.
