+++ scripts/env.lua
@@ -2,10 +2,19 @@
 package.path = package.path .. ";./scripts/?.lua"
 package.path = package.path .. ";./scripts/common/?.lua"
 package.path = package.path .. ";./scripts/quests/?.lua"
+package.path = package.path .. ";./scripts/quests/aeschburh/?.lua"
+package.path = package.path .. ";./scripts/quests/atuaelar/?.lua"
 package.path = package.path .. ";./scripts/quests/carcassia/?.lua"
-package.pack = package.path .. ";./scripts/quests/isen_dun/?.lua"
+package.path = package.path .. ";./scripts/quests/fae_caravan/?.lua"
+package.path = package.path .. ";./scripts/quests/gnordvar/?.lua"
+package.path = package.path .. ";./scripts/quests/isen_dun/?.lua"
+package.path = package.path .. ";./scripts/quests/last_islands/?.lua"
+package.path = package.path .. ";./scripts/quests/rowan_head/?.lua"
 package.path = package.path .. ";./scripts/quests/sithrin/?.lua"
+package.path = package.path .. ";./scripts/quests/stoneheim/?.lua"
 package.path = package.path .. ";./scripts/quests/telari/?.lua"
+package.path = package.path .. ";./scripts/quests/whalings_end/?.lua"
+package.path = package.path .. ";./scripts/quests/yew/?.lua"
 package.path = package.path .. ";./scripts/level/?.lua"
 package.path = package.path .. ";./scripts/areas/?.lua"
 package.path = package.path .. ";./scripts/areas/events/?.lua"
