+++ scripts/quests/atuaelar/kirin.lua
