+++ scripts/quests/atuaelar/tir.lua
