+++ scripts/quests/fae_caravan/eakauri.lua
