+++ scripts/quests/gnordvar/christof.lua
