+++ scripts/quests/last_islands/amaurosis.lua
