+++ scripts/quests/last_islands/bagra.lua
