+++ scripts/quests/rowan_head/exile.lua
