+++ scripts/quests/rowan_head/maisy.lua
