+++ scripts/quests/sithrin/altus.lua
