+++ scripts/quests/sithrin/cornel.lua
