+++ scripts/quests/stoneheim/gildi.lua
