+++ scripts/quests/stoneheim/skaal.lua
