+++ scripts/quests/whalings_end/gnomish_seatrader.lua
